fix: delete prodak

+++ src/controllers/productController.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Product from "../models/Product.js";
 
 //GET
@@ -26,7 +27,12 @@ export const createProducts = async(req, res) => {
 // DELETE
 export const deleteProduct = async(req, res) => {
     try {
-        const deletedProduct = Product.findByIdAndDelete(req.params.id);
+
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({message: "Invalid ID Format"});
+        }
+
+        const deletedProduct = await Product.findByIdAndDelete(req.params.id);
 
         if (!deletedProduct) {
             return res.status(404).json({message: "Product Not Found"});
